Avoid g++ warnings.

+++ src/output.cc
@@ -1299,8 +1299,8 @@ Output::output_lookup_array () const
 
           if (option[DEBUG])
             fprintf (stderr,
-                     "dup_ptr[%d]: hash_value = %d, index = %d, count = %d\n",
-                     dup_ptr - duplicates,
+                     "dup_ptr[%lu]: hash_value = %d, index = %d, count = %d\n",
+                     static_cast<unsigned long>(dup_ptr - duplicates),
                      dup_ptr->hash_value, dup_ptr->index, dup_ptr->count);
 
           int i;
